+++ backend/src/app.module.ts
@@ -4,15 +4,15 @@ import { typeORMConfig } from './configs/typeorm.config';
 import { AuthModule } from './auth/auth.module';
 import { UsersModule } from './users/users.module';
 import { ConfigModule } from '@nestjs/config';
-
+import { forwardRef } from '@nestjs/common';
 @Module({
   imports: [
     ConfigModule.forRoot({
       isGlobal: true,
     }),
     TypeOrmModule.forRoot(typeORMConfig),
-    AuthModule,
     UsersModule,
+    AuthModule,
   ],
 })
 export class AppModule {}
