+++ backend/src/auth/auth.service.ts
@@ -1,12 +1,13 @@
 import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
-import { User } from 'src/users/user.entity';
 import * as config from 'config';
+import { User } from 'src/users/user.entity';
 
 const jwtConfig = config.get('jwt');
 const refreshConfig = config.get('refresh');
+
 @Injectable()
-export class JwtAuthService {
+export class AuthService {
   constructor(private jwtService: JwtService) {}
 
   getAccessToken(user: User) {
