--- backend/src/auth/google/google-oauth.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UsersModule } from 'src/users/users.module';
-import { JwtAuthModule } from '../jwt/jwt-auth.module';
-import { GoogleOauthController } from './google-oauth.controller';
-import { GoogleOauthStrategy } from './google-oauth.strategy';
-
-@Module({
-  imports: [UsersModule, JwtAuthModule],
-  controllers: [GoogleOauthController],
-  providers: [GoogleOauthStrategy],
-})
-export class GoogleOauthModule {}
