+++ backend/src/auth/strategies/jwt-auth.strategy.ts
