+++ backend/src/users/users.controller.ts
@@ -1,8 +1,8 @@
 import { Controller, Get, Logger, Req, Res, UseGuards, Inject, forwardRef } from '@nestjs/common';
 import { Request, Response } from 'express';
-import { JwtAuthGuard } from 'src/auth/jwt/jwt-auth.guard';
-import { JwtAuthService } from 'src/auth/jwt/jwt-auth.service';
-import { JwtRefreshGuard } from 'src/auth/jwt/jwt-refresh.guard';
+import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
+import { JwtRefreshGuard } from 'src/auth/guards/jwt-refresh.guard';
+import { AuthService } from 'src/auth/auth.service';
 import { ResponseUserDto } from './dto/response-user.dto';
 import { User } from './user.entity';
 import { UsersService } from './users.service';
@@ -10,7 +10,7 @@ import { UsersService } from './users.service';
 @Controller('users')
 export class UsersController {
   private logger = new Logger('UserController');
-  constructor(private readonly userService: UsersService, private jwtAuthService: JwtAuthService) {}
+  constructor(private readonly userService: UsersService, private authService: AuthService) {}
 
   @Get()
   @UseGuards(JwtAuthGuard)
@@ -24,7 +24,7 @@ export class UsersController {
   @UseGuards(JwtRefreshGuard)
   async refresh(@Req() req, @Res({ passthrough: true }) res: Response): Promise<ResponseUserDto> {
     const user = req.user;
-    const { accessToken, ...accessOption } = this.jwtAuthService.getAccessToken(user);
+    const { accessToken, ...accessOption } = this.authService.getAccessToken(user);
     res.cookie('jwt', accessToken, accessOption);
     return await this.userService.getUserInfo(user);
   }
